requirements.txt
Flask==1.1.2
numpy==1.18.5
json5==0.9.5
requests==2.24.0
Pillow==8.0.1


